+++ src/main/java/com/mapshot/api/domain/admin/notice/AdminNoticeRepository.java
@@ -1,0 +1,7 @@
+package com.mapshot.api.domain.admin.notice;
+
+import com.mapshot.api.domain.notice.NoticeEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminNoticeRepository extends JpaRepository<NoticeEntity, Long> {
+}
