+++ src/main/java/com/mapshot/api/domain/admin/user/AdminUserService.java
@@ -1,9 +1,5 @@
 package com.mapshot.api.domain.admin.user;
 
-import com.mapshot.api.domain.community.post.PostRepository;
-import com.mapshot.api.domain.notice.NoticeEntity;
-import com.mapshot.api.domain.notice.NoticeRepository;
-import com.mapshot.api.domain.notice.NoticeType;
 import com.mapshot.api.infra.auth.Validation;
 import com.mapshot.api.infra.encrypt.EncryptUtil;
 import com.mapshot.api.infra.exception.ApiException;
@@ -18,9 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class AdminUserService {
 
     private final AdminUserRepository adminUserRepository;
-    private final PostRepository postRepository;
     private final Validation adminValidation;
-    private final NoticeRepository noticeRepository;
 
     @Transactional(readOnly = true)
     public void validateUser(String nickname, String password) {
@@ -28,42 +22,6 @@ public class AdminUserService {
 
         adminUserRepository.findByNicknameAndPassword(nickname, password)
                 .orElseThrow(() -> new ApiException(ErrorCode.NO_SUCH_USER));
-    }
-
-    @Transactional
-    public void deletePost(long id) {
-        postRepository.deleteById(id);
-    }
-
-    @Transactional
-    public long saveNotice(NoticeType type, String title, String content) {
-
-        return noticeRepository.save(NoticeEntity.builder()
-                        .noticeType(type)
-                        .title(title)
-                        .content(content)
-                        .build())
-                .getId();
-    }
-
-    @Transactional
-    public long modifyNotice(long id, NoticeType type, String title, String content) {
-
-        NoticeEntity noticeEntity = noticeRepository.findById(id)
-                .orElseThrow(() -> new ApiException(ErrorCode.NO_SUCH_NOTICE));
-
-        noticeEntity.update(title, type, content);
-
-        return noticeRepository.save(noticeEntity).getId();
-    }
-
-
-    @Transactional
-    public void deleteNotice(long id) {
-        NoticeEntity noticeEntity = noticeRepository.findById(id)
-                .orElseThrow(() -> new ApiException(ErrorCode.NO_SUCH_NOTICE));
-
-        noticeRepository.delete(noticeEntity);
     }
