--- src/main/java/com/mapshot/api/presentation/admin/AdminController.java
@@ -1,80 +1,0 @@
-package com.mapshot.api.presentation.admin;
-
-import com.mapshot.api.domain.admin.user.AdminUserService;
-import com.mapshot.api.domain.notice.NoticeType;
-import com.mapshot.api.infra.auth.annotation.PreAuth;
-import com.mapshot.api.infra.auth.enums.Accessible;
-import com.mapshot.api.presentation.notice.NoticeRegistrationRequest;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.PositiveOrZero;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/admin")
-@CrossOrigin(originPatterns = {"https://*.kmapshot.com", "https://kmapshot.com"})
-@Validated
-public class AdminController {
-
-    private final AdminUserService adminUserService;
-
-    @PreAuth(Accessible.EVERYONE)
-    @PostMapping("/login")
-    public ResponseEntity<Void> login(@RequestBody AdminRequest request) {
-        adminUserService.validateUser(request.getNickname(), request.getPassword());
-        HttpHeaders authHeader = adminUserService.getAuthHeader();
-
-        return ResponseEntity.ok()
-                .headers(authHeader)
-                .build();
-    }
-
-    @PreAuth(Accessible.ADMIN)
-    @PostMapping("/auth/refresh")
-    public ResponseEntity<Void> refreshAuth() {
-        HttpHeaders authHeader = adminUserService.getAuthHeader();
-
-        return ResponseEntity.ok()
-                .headers(authHeader)
-                .build();
-    }
-
-    @PreAuth(Accessible.ADMIN)
-    @PostMapping("/notice/register")
-    public ResponseEntity<Void> registerNotice(@RequestBody NoticeRegistrationRequest request) {
-        adminUserService.saveNotice(NoticeType.valueOf(request.getNoticeType()), request.getTitle(), request.getContent());
-
-        return ResponseEntity.ok().build();
-    }
-
-    @PreAuth(Accessible.ADMIN)
-    @GetMapping("/notice/delete/{noticeNumber}")
-    public ResponseEntity<Void> deleteNotice(@PositiveOrZero @PathVariable(value = "noticeNumber") long noticeNumber) {
-        adminUserService.deleteNotice(noticeNumber);
-
-        return ResponseEntity.ok().build();
-    }
-
-
-    @PreAuth(Accessible.ADMIN)
-    @PostMapping("/notice/modify/{noticeNumber}")
-    public ResponseEntity<Void> modifyNotice(@PositiveOrZero @PathVariable(value = "noticeNumber") long noticeNumber,
-                                             @RequestBody NoticeRegistrationRequest request) {
-        adminUserService.modifyNotice(noticeNumber, NoticeType.valueOf(request.getNoticeType()), request.getTitle(), request.getContent());
-
-        return ResponseEntity.ok().build();
-    }
-
-
-    @PreAuth(Accessible.ADMIN)
-    @GetMapping("/post/delete/{postNumber}")
-    public ResponseEntity<Void> deletePost(@Positive @PathVariable(value = "postNumber") long postNumber) {
-        adminUserService.deletePost(postNumber);
-
-        return ResponseEntity.ok().build();
-    }
-}
